+++ apps/web/lib/statistics.ts
